CThemes: visualize migrated theme after possible update
After an update this entry will be only shown if theme name is a new
configure option and theme name is still unknown.
This entry visualizes the state of current migrated theme. The old theme
settings will be kept.
If required users should save the theme as a new userdefined theme or
select a given theme.