Imageinfo: prepare functionality to show privacy policy
Note:
Contents of policy files are not available with this commit. This
should be done related to own conditions by provider himself.
Providers are required by law to do this for their users as well.

Files must be installed in:

	installdir = $(DATADIR)/neutrino/policy

Providers should modify their buildsystem to install these files.
Filename sheme is
	<locale>.policy. eg 'english.policy'

If you want to add a policy file, take a look into sample file
	locale.policy.sample