helpers.cpp: fix narrowing conversion
switch expression is 'int', but f_type gets stored as 'long unsigned int' on 32-bit machines since it may not fit otherwise. Recent compilers diagnose this and bail out with:

../../../git/src/system/helpers.cpp:330:9: error: narrowing conversion of '2240043254' from 'long unsigned int' to 'int' [-Wnarrowing]

casting the switch expression to long unsigned int seems to be the solution.

in short: this fixes the build for 32-bit machines with gcc ≥ 10.x

Signed-off-by: Markus Volk <f_l_k@t-online.de>