system/debug.h: try for more thread save implementation
Replacing printf() with fprintf() in this code section aims to
prevent unpredictable behavior in multi-threaded environments that
could be caused by printf()'s non-atomic nature. The change is
intended to enhance output consistency and prevent potential race conditions.
Additionally, it is important to note that many other plain calls to printf()
in the Neutrino code could also cause issues and require further investigation.
Whether this change will have a positive impact remains to be seen.